move things related to server to serverv2 core
